feat: add fallback if bot fails to load the show article
In read_curr_show_button, add a try exception handle, which will catch
error in case the article fails to load by the bot in embed due to
any reason, then Send the article's url instead to the user.

Issues due to which article may not load right now will be fixed in a
future commit.